mab/requirements.txt
# Bernoulli MAB + HF LLM/scoring (run from `mab/`).
numpy>=1.26.0
matplotlib>=3.9.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.26.0
python-dotenv>=1.0.0
